[RDoc-1223] refactoring
