+++ Raven.Documentation.Parser/DocumentationParser.cs
@@ -1,22 +1,29 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Raven.Documentation.Parser.Compilation;
+using Raven.Documentation.Parser.Compilation.DocumentationDirectory;
+using Raven.Documentation.Parser.Compilation.ToC;
 using Raven.Documentation.Parser.Data;
 
 namespace Raven.Documentation.Parser
 {
     public class DocumentationParser : ParserBase<DocumentationPage>
     {
-        private readonly DocumentationDirectoryCompiler _directoryCompiler;
         private readonly TableOfContentsCompiler _tableOfContentsCompiler;
+        private readonly TocBasedPagesCompiler _tocBasedPagesCompiler;
+        private readonly IndexPagesCompiler _indexPagesCompiler;
 
         public DocumentationParser(ParserOptions options, IProvideGitFileInformation repoAnalyzer)
             : base(options)
         {
             var documentCompiler = new DocumentCompiler(Parser, options, repoAnalyzer);
-            _directoryCompiler = new DocumentationDirectoryCompiler(documentCompiler, options);
             _tableOfContentsCompiler = new TableOfContentsCompiler(options);
+            _tocBasedPagesCompiler = new TocBasedPagesCompiler(documentCompiler, Options);
+            _indexPagesCompiler = new IndexPagesCompiler(documentCompiler, Options);
         }
+
+        private bool CompileAllVersions => Options.VersionsToParse.Count == 0;
 
         public override ParserOutput Parse()
         {
@@ -32,7 +39,7 @@ namespace Raven.Documentation.Parser
 
         private IEnumerable<TableOfContents> GenerateTableOfContents()
         {
-            var compilationContext = new DocumentationCompilation.Context();
+            var compilationContext = new CompilationUtils.Context();
 
             var documentationDirectories = Directory.GetDirectories(Options.PathToDocumentationDirectory);
 
@@ -41,48 +48,8 @@ namespace Raven.Documentation.Parser
                 .SelectMany(directory => _tableOfContentsCompiler.GenerateTableOfContents(directory, compilationContext))
                 .ToList();
 
-            var duplicatedTocs = tableOfContents.GroupBy(x => new { x.Version, x.Category })
-                .Where(g => g.Count() > 1)
-                .ToList();
-
-            foreach (var duplicatedTocGroup in duplicatedTocs)
-            {
-                var mergedToc = MergeTocs(duplicatedTocGroup.ToList());
-                ReplaceToc(mergedToc, tableOfContents);
-            }
-
+            TableOfContentsMerger.Merge(tableOfContents);
             return tableOfContents;
-        }
-
-        private TableOfContents MergeTocs(List<TableOfContents> tocs)
-        {
-            var toc = tocs.First();
-
-            for (var i = 1; i < tocs.Count; i++)
-            {
-                MergeTocItems(tocs[i].Items, toc.Items);
-            }
-
-            return toc;
-        }
-
-        private void MergeTocItems(List<TableOfContents.TableOfContentsItem> source, List<TableOfContents.TableOfContentsItem> destination)
-        {
-            foreach (var sourceItem in source)
-            {
-                var matchedItem = destination.SingleOrDefault(x => x.Key == sourceItem.Key);
-
-                if (matchedItem == null)
-                    destination.Add(sourceItem);
-                else
-                    MergeTocItems(sourceItem.Items, matchedItem.Items);
-            }
-        }
-
-        private void ReplaceToc(TableOfContents mergedToc, List<TableOfContents> tableOfContents)
-        {
-            tableOfContents.RemoveAll(x => x.Version == mergedToc.Version && x.Category == mergedToc.Category);
-            tableOfContents.Add(mergedToc);
         }
 
         private IEnumerable<DocumentationPage> GenerateDocumentationPages(IEnumerable<TableOfContents> tocs)
@@ -100,15 +67,15 @@ namespace Raven.Documentation.Parser
 
             return documentationDirectories
                 .Select(documentationDirectory => new DirectoryInfo(documentationDirectory))
-                .Where(documentationDirectory => Options.VersionsToParse.Count == 0 || Options.VersionsToParse.Contains(documentationDirectory.Name))
-                .SelectMany(_directoryCompiler.CompileIndexFiles);
+                .Where(documentationDirectory => CompileAllVersions || Options.VersionsToParse.Contains(documentationDirectory.Name))
+                .SelectMany(_indexPagesCompiler.Compile);
         }
 
         private IEnumerable<DocumentationPage> GenerateDocumentationPagesFromToc(IEnumerable<TableOfContents> tocs)
         {
             return tocs
-                .Where(toc => Options.VersionsToParse.Count == 0 || Options.VersionsToParse.Contains(toc.Version))
-                .SelectMany(_directoryCompiler.Compile);
+                .Where(toc => CompileAllVersions || Options.VersionsToParse.Contains(toc.Version))
+                .SelectMany(_tocBasedPagesCompiler.Compile);
         }
     }
 }
